# packages/server/api/v1.swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for Chalarí",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/memos": {
            "summary": "Handling memos.",
            "get": {
                "tags": ["Memos"],
                "parameters": [
                    {  
                        "name": "search",
                        "in": "query",
                        "description": "Filter the memos via the title (case-insensitive). Don't set or set an empty string to omit.",
                        "schema": {
                            "type": "string",
                            "example": "Memo",
                            "default": ""
                        }
                    },
                    {  
                        "name": "category",
                        "in": "query",
                        "description": "Filter the memos via the category. Don't set or set an empty string to omit.",
                        "schema": {
                            "type": "string",
                            "example": "Ideas",
                            "default": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieved the memos successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrieved the memos successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "qwg9efb-1r2egqw87e9f-1238rg2f-12f7h2"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Ideas"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "My new memo"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "This is my new memo !!"
                                                    },
                                                    "isPinned": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "updatedAt": {
                                                        "type": "string",
                                                        "format": "ISO-8601",
                                                        "example": "2024-08-15T12:30:00.000Z"
                                                    }
                                                }
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Memo with given id not found.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Memo with given id not found."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong whilst retrieving the memos."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["Memos"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "category": {
                                        "type": "string",
                                        "example": "Ideas",
                                        "default": null
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "My new memo"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "The content of my newly created memo !",
                                        "default": ""
                                    },
                                    "isPinned": {
                                        "type": "boolean",
                                        "example": false,
                                        "default": false
                                    }
                                },
                                "required": ["title"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Created the memo successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Created the memo successfully."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "qwg9efb-1r2egqw87e9f-1238rg2f-12f7h2"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Ideas"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "The content of my modified memo !"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "This is my new memo !!"
                                                },
                                                "isPinned": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "updatedAt": {
                                                    "type": "string",
                                                    "format": "ISO-8601",
                                                    "example": "2024-08-15T12:30:00.000Z"
                                                }
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The body format doesn't match the expected format.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad request format."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong whilst creating the memo."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}